Move storage.mts out of "privileged" dir
- Move storage.mts into src/modules out of src/modules/privileged, to
  indicate that it is available to non-privileged (i.e. content) scripts
