+++ src/modules/page/build.mts
@@ -1,5 +1,5 @@
 import * as Manifest from "/dist/modules/manifest.mjs";
-import { Bank, Config } from "/dist/modules/privileged/storage.mjs";
+import { Bank, Config } from "../storage.mjs";
 import type { MatchTerm } from "/dist/modules/match-term.mjs";
 import { getIdSequential } from "/dist/modules/common.mjs";
 import * as EmailJS from "/lib/email.min.js";
