+++ src/pages/options-new.mts
@@ -5,8 +5,8 @@ import {
 } from "/dist/modules/page/build.mjs";
 import type {
 	StoreImmediate, StoreList, StoreListInterface, ConfigValues, ConfigKey,
-} from "/dist/modules/privileged/storage.mjs";
-import { StoreType, Config } from "/dist/modules/privileged/storage.mjs";
+} from "../modules/storage.mjs";
+import { StoreType, Config } from "../modules/storage.mjs";
 import { compatibility } from "/dist/modules/common.mjs";
 
 const getControlOptionTemp = <ConfigK extends ConfigKey>(
