+++ src/pages/options.mts
@@ -2,8 +2,8 @@
 //import { isWindowInFrame } from "/dist/modules/page/build.mjs";
 import type {
 	StoreImmediate, StoreList, StoreListInterface, ConfigValues, ConfigKey,
-} from "/dist/modules/privileged/storage.mjs";
-import { StoreType, Config } from "/dist/modules/privileged/storage.mjs";
+} from "../modules/storage.mjs";
+import { StoreType, Config } from "../modules/storage.mjs";
 import { compatibility, getIdSequential } from "/dist/modules/common.mjs";
 
 const isWindowInFrame = () => (
