+++ src/pages/startpage.mts
@@ -1,6 +1,6 @@
 import * as Manifest from "/dist/modules/manifest.mjs";
 import { type Page, loadPage } from "/dist/modules/page/build.mjs";
-import { Config } from "/dist/modules/privileged/storage.mjs";
+import { Config } from "../modules/storage.mjs";
 
 /**
  * Loads the startpage content into the page.
